Fixed a issue with the output wires of module instantiation.
When istantiating modules, the output wire size was determined by evaluating an expression.
Even if the expression contained dependencies on parameters only visible in the instantiated module, no attempt to resolve them in the instantiating module, so iverilog reported unfound names error.
Unfortunately, a clean solution didn't exist - I did a hack, and I'll think harder about this issue for a next version.