Fixed empty attribute list instantiation issue.
If a module with no attributes was instantiated, our compiler would generate for it an empty attribute list. This cause problems with XST thinking the module should have at least one parameter.
Fixed this, so instantations for modules or fsms with no attributes do not generate the empty parameter list.